requirements.txt
# InfraVision AI - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Framework & Web
Flask==2.3.3
Flask-CORS==4.0.0
streamlit==1.28.1
Werkzeug==2.3.7

# Data Science & Analytics
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2
scikit-learn==1.3.0
statsmodels==0.14.0

# Computer Vision & Image Processing
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
Pillow==10.0.0
imageio==2.32.0
scikit-image==0.21.0

# Object Detection (YOLOv8)
ultralytics==8.0.189
torch==2.0.1
torchvision==0.15.2
onnx==1.14.0

# Deep Learning & Neural Networks
tensorflow==2.13.0
keras==2.13.1

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# PDF & Reporting
reportlab==4.0.4
PyPDF2==3.0.1
python-pptx==0.6.21

# API & HTTP
requests==2.31.0
httpx==0.24.1

# Database & Storage
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7

# Image Analysis & Processing
Pillow-SIMD==9.2.0
imutils==0.5.4
trimesh==3.21.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7
PyYAML==6.0.1

# Monitoring & Logging
loguru==0.7.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
pylint==2.17.5
mypy==1.4.1
jupyter==1.0.0
ipython==8.14.0

